requirements.txt
opencv-python==4.9.0.80
librosa==0.10.1
scipy==1.13.0
spafe==0.3.2
praat-parselmouth==0.4.3
noisereduce==3.0.2
python-dotenv==1.0.1
pydub
pyannote.audio
pickles==0.1.1
ffmpeg-python
PyQt5==5.15.10
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.2.1 
torchvision==0.17.1 
torchaudio==2.2.1 
transformers==4.40.1
datasets==2.19.0